Add database and crude view
This commit adds a PostgreSQL database. The migration is crude and
basically just creates a new table.
The database must be accessible on localhost or through the
`DATABASE_URL` environment variable.
The test to see if the database exists (and was seeded during the
migration) is visiting the root `localhost:3000`. The page will show the
first row in the `todos` table which should be populated.