Add JSON index endpoint
This is really just a bit of a wrapper that isn't much different than
its web version.
The views have now been separated into `http` and `json` files. I am not
sure that this is the right route, but hopefully continued development
will show the best way to organize these files.
Also added a bit of logic around what gets published. The HTTP and JSON
endpoints shouldn't really care about the `created_at` timestamps. This
was solved on the HTTP side through the formatter, but that shouldn't
necessarily be a presentation concernt. This commit adds a function on
the model that will only send the necessary information out to the
views. That way they can stay dumber, which seems like the right choice.