Add endpoint to create new todos
This was tricky as compojure default CSRF protection and JSON request
parsing were tricky. My solutions are pbviously not ready for
production, but they work for adding records to the database.
A few changes I would like to make include validating request parameters
and dealing more gracefully with errors in writing to the database.