Add show endpoint and clean up README
I figured that a show endpoint could be useful. I chose not to return
the specific todo being created or updated in the response to those
requests because the client should have that data already and can choose
to paint the page with it or not. Of course, it would be easy to add
that in if necessary.
Other changes in this commit are basically cleanup in the README and
ordering of functions for consistency.